\name{geneLevelData}
\alias{geneLevelData}
\docType{data}
\title{
  Yeast gene-level counts from: Lee et al. PloS Genetics 2008.
}
\description{
  Gene counts for an Illumina sequencing experiment. These counts
  represent summarizations over the data in the yeastAligned object in
  this package.
}
\usage{data(geneLevelData)}
\format{
}
\details{
}
\source{
}
\references{
}
\examples{
data(geneLevelData)
## maybe str(geneLevelData) ; plot(geneLevelData) ...
}
\keyword{datasets}
