\name{yeastAligned}
\alias{yeastAligned}
\docType{data}
\title{
  AlignedRead list
}
\description{
  A four element list of AlignedRead objects from the ShortRead
  package from 4 yeast lanes of a Solexa experiment:
  Lee et al. PloS Genetics 2008.

  The data represents a random subsample of reads from 4 lanes (2
  wild-type, 2 mutant lanes)
}
\usage{data(yeastAligned)}
\format{
}
\details{
}
\source{
}
\references{
}
\examples{
data(yeastAligned)
## maybe str(yeastAligned) ; plot(yeastAligned) ...
}
\keyword{datasets}
