\name{yeastAnno}
\alias{yeastAnno}
\docType{data}
\title{
  Yeast Annotation
}
\description{
  Annotation for S. cerevisiae downloaded from sgd
  (http://www.yeastgenome.org/) in the fall of 2009 and
  post-processed. This file should not be used for analyses purposes,
  but rather demonstrates some functionality of various packages. 

}
\usage{data(yeastAnno)}
\format{
}
\details{
}
\source{
}
\references{
}
\examples{
data(yeastAnno)
}
\keyword{datasets}
