\name{geneLevelData}
\alias{geneLevelData}
\docType{data}
\title{
  Yeast gene-level counts from: Lee et al. PloS Genetics 2008.
}
\description{
  Gene counts for an Illumina sequencing experiment. These counts
  represent summarizations over the data in the \code{yeastAligned}
  object using the annotation from the \code{yeastAnno} object, both
  from this package.
}
\usage{data(geneLevelData)}
\source{
  A citation for the dataset may be generated using
  \code{citation("yeastRNASeq")}. 
}
\seealso{
  The package vignette.
}
\keyword{datasets}
