\name{yeastAnno}
\alias{yeastAnno}
\docType{data}
\title{
  Yeast Annotation
}
\description{
  Annotation for S. cerevisiae downloaded from Ensembl using the
  \pkg{biomaRt} package in the fall of 2009 and
  post-processed. This file should not be used for analyses purposes,
  but rather demonstrates some functionality of various packages. 
}
\usage{data(geneLevelData)}
\source{
  A citation for the dataset may be generated using
  \code{citation("yeastRNASeq")}. 
}
\seealso{
  The package vignette.
}
\keyword{datasets}
